Move huawei_lte base entity to separate module (#126098)
