+++ homeassistant/components/huawei_lte/button.py
@@ -16,8 +16,8 @@ from homeassistant.const import EntityCategory
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers import entity_platform
 
-from . import HuaweiLteBaseEntityWithDevice
 from .const import DOMAIN
+from .entity import HuaweiLteBaseEntityWithDevice
 
 _LOGGER = logging.getLogger(__name__)
