+++ homeassistant/components/huawei_lte/switch.py
@@ -15,12 +15,12 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity import Entity
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import HuaweiLteBaseEntityWithDevice
 from .const import (
     DOMAIN,
     KEY_DIALUP_MOBILE_DATASWITCH,
     KEY_WLAN_WIFI_GUEST_NETWORK_SWITCH,
 )
+from .entity import HuaweiLteBaseEntityWithDevice
 
 _LOGGER = logging.getLogger(__name__)
